requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
requests==2.31.0
pandas==2.1.3
numpy==1.26.2
sqlalchemy==2.0.23
scipy==1.11.4
reportlab==4.0.7
python-dotenv==1.0.0
Werkzeug==3.0.1
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0
openpyxl==3.1.2
xlsxwriter==3.1.9
Flask-Mail==0.9.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
APScheduler==3.10.4
celery==5.3.4
redis==5.0.1
Flask-Limiter==3.5.0